refactor(#185): DTO에 of/from 패턴 적용 및 변환 책임 이동
