+++ src/main/kotlin/team/mozu/dsm/adapter/in/team/dto/response/StockResponse.kt
@@ -1,5 +1,7 @@
 package team.mozu.dsm.adapter.`in`.team.dto.response
 
+import team.mozu.dsm.domain.lesson.model.LessonItem
+import team.mozu.dsm.domain.team.model.Stock
 import java.util.UUID
 
 data class StockResponse(
@@ -13,4 +15,34 @@ data class StockResponse(
     val valuationMoney: Long,
     val valProfit: Long,
     val profitNum: Double
-)
+) {
+    companion object {
+        fun of(stock: Stock, lessonItem: LessonItem, currentRound: Int, previousRound: Int): StockResponse {
+            val currentPrice = lessonItem.getPriceByRound(currentRound) ?: lessonItem.endPrice
+            val previousPrice = lessonItem.getPriceByRound(previousRound) ?: lessonItem.endPrice
+
+            val currentValuation = currentPrice * stock.quantity
+            val previousValuation = previousPrice * stock.quantity
+            val valProfit = currentValuation - previousValuation
+
+            val profitNum = if (previousValuation > 0) {
+                (valProfit.toDouble() / previousValuation.toDouble()) * 100
+            } else {
+                0.0
+            }
+
+            return StockResponse(
+                id = stock.id!!,
+                itemId = stock.itemId,
+                itemName = stock.itemName,
+                avgPurchasePrice = stock.avgPurchasePrice,
+                quantity = stock.quantity,
+                totalMoney = stock.buyMoney,
+                nowMoney = currentPrice,
+                valuationMoney = currentValuation,
+                valProfit = valProfit,
+                profitNum = profitNum
+            )
+        }
+    }
+}
