+++ src/main/kotlin/team/mozu/dsm/adapter/in/team/dto/response/TeamRankResponse.kt
@@ -1,5 +1,6 @@
 package team.mozu.dsm.adapter.`in`.team.dto.response
 
+import team.mozu.dsm.domain.team.model.Team
 import java.util.UUID
 
 data class TeamRankResponse(
@@ -8,4 +9,16 @@ data class TeamRankResponse(
     val schoolName: String,
     val totalMoney: Long,
     val isMyTeam: Boolean
-)
+) {
+    companion object {
+        fun of(team: Team, totalMoney: Long, myTeamId: UUID): TeamRankResponse {
+            return TeamRankResponse(
+                id = team.id!!,
+                teamName = team.teamName,
+                schoolName = team.schoolName,
+                totalMoney = totalMoney,
+                isMyTeam = team.id == myTeamId
+            )
+        }
+    }
+}
