+++ src/main/kotlin/team/mozu/dsm/application/service/team/QueryHoldStocksService.kt
@@ -23,13 +23,11 @@ class QueryHoldStocksService(
     @Transactional(readOnly = true)
     override fun execute(teamId: UUID): List<StockResponse> {
         return try {
-            val team = queryTeamPort.findById(teamId) ?: run {
-                throw TeamNotFoundException
-            }
+            val team = queryTeamPort.findById(teamId)
+                ?: throw TeamNotFoundException
 
-            val lesson = queryLessonPort.findByLessonNum(team.lessonNum) ?: run {
-                throw LessonNotFoundException
-            }
+            val lesson = queryLessonPort.findByLessonNum(team.lessonNum)
+                ?: throw LessonNotFoundException
 
             val stocks = queryStockPort.findAllByTeamId(teamId)
                 .filter { it.id != null && it.quantity > 0 }
@@ -47,32 +45,7 @@ class QueryHoldStocksService(
 
             stocks.mapNotNull { stock ->
                 val lessonItem = lessonItemMap[stock.itemId] ?: return@mapNotNull null
-
-                val currentPrice = lessonItem.getPriceByRound(currentRound) ?: lessonItem.endPrice
-                val previousPrice = lessonItem.getPriceByRound(previousRound) ?: lessonItem.endPrice
-
-                val currentValuation = currentPrice * stock.quantity
-                val previousValuation = previousPrice * stock.quantity
-                val valProfit = currentValuation - previousValuation
-
-                val profitNum = if (previousValuation > 0) {
-                    (valProfit.toDouble() / previousValuation.toDouble()) * 100
-                } else {
-                    0.0
-                }
-
-                StockResponse(
-                    id = stock.id!!,
-                    itemId = stock.itemId,
-                    itemName = stock.itemName,
-                    avgPurchasePrice = stock.avgPurchasePrice,
-                    quantity = stock.quantity,
-                    totalMoney = stock.buyMoney,
-                    nowMoney = currentPrice,
-                    valuationMoney = currentValuation,
-                    valProfit = valProfit,
-                    profitNum = profitNum
-                )
+                StockResponse.of(stock, lessonItem, currentRound, previousRound)
             }
         } catch (e: Exception) {
             e.printStackTrace()
