+++ src/main/kotlin/team/mozu/dsm/application/service/team/QueryTeamDetailService.kt
@@ -42,7 +42,6 @@ class QueryTeamDetailService(
         val currentStockValuation = stocks.sumOf { stock ->
             val lessonItem = lessonItemMap[stock.itemId]
                 ?: throw LessonItemNotFoundException
-
             val currentPrice = lessonItem.getPriceByRound(currentRound) ?: lessonItem.endPrice
             currentPrice * stock.quantity
         }
@@ -51,42 +50,10 @@ class QueryTeamDetailService(
         val previousStockValuation = stocks.sumOf { stock ->
             val lessonItem = lessonItemMap[stock.itemId]
                 ?: throw LessonItemNotFoundException
-
             val previousPrice = lessonItem.getPriceByRound(previousRound) ?: lessonItem.endPrice
             previousPrice * stock.quantity
         }
 
-        // 실제 총 자산 = 현금 + 주식 평가액 (현재 차수 기준)
-        val actualTotalMoney = team.cashMoney + currentStockValuation
-
-        // 이전 차수 기준 총 자산 = 현금 + 이전 차수 주식 평가액
-        val previousTotalMoney = team.cashMoney + previousStockValuation
-
-        // 총 자산 증감률 (이전 vs 현재 차수)
-        val currentTotalValProfit = actualTotalMoney - previousTotalMoney
-
-        val profitNum = if (previousTotalMoney > 0) {
-            (currentTotalValProfit.toDouble() / previousTotalMoney.toDouble()) * 100
-        } else {
-            0.0
-        }
-
-        val formattedProfitNum = when {
-            profitNum > 0 -> "+%.2f%%".format(profitNum)
-            else -> "%.2f%%".format(profitNum)
-        }
-
-        return TeamDetailResponse(
-            id = teamId,
-            teamName = team.teamName,
-            baseMoney = lesson.baseMoney,
-            totalMoney = actualTotalMoney,
-            cashMoney = team.cashMoney,
-            valuationMoney = currentStockValuation,
-            curInvRound = currentRound,
-            maxInvRound = lesson.maxInvRound,
-            valProfit = currentTotalValProfit,
-            profitNum = formattedProfitNum
-        )
+        return TeamDetailResponse.of(team, lesson, currentStockValuation, previousStockValuation)
     }
 }
